Add TLS configuration details to README and update dependencies
This commit introduces Transport Layer Security (TLS) configuration details for both HTTP endpoints and Feign HTTP clients in the README file. It also includes necessary updates to the application's Java implementation, including addition of new security dependencies in pom.xml, and initiation of the BouncyCastleFipsProvider in the main application file. The edge-common-spring version was updated as well, alongside some modifications in the application-test.yml file.